Finish disk restore feature, improve unit tests, fix module pathing, performance and return type changes, improved error handling
